requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.1
Pillow==7.0.0
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
